Replace custom CLI parsing with argparse (#10)
* Switch from custom CLI parsing to argparse

The code has been refactored to always enter through the `run()`
function, reducing the need for duplicate code. This allowed the removal
of two custom functions, which are now unused:

- `showHelp()`
- `printexit()`

One thing left to fix is backwards compatibility. The help is no brought
up through '--help' or '-h', not '-H' as it was previously.

* Break up a really long line of text

* Change the help short flag from '-h' to '-H'

* Add help message to '--config' flag